refactor: myInfo 쿼리 구조 정리 및 참조 경로 개선 (#103)
- 내 정보 관련 훅을 myInfo 폴더로 분리
- 마이페이지 내 정보 화면에서 변경된 쿼리 경로로 참조 수정
- 로딩 UI 감싸는 레이아웃을 보완해 화면 정렬 개선
